refactor(#176): move VersionedStorage interface to a file
