+++ src/types/utils.ts
@@ -26,42 +26,6 @@ export type MutableValue<Name extends string, Type> = {
 export type RecordValues<T extends Record<any, any>> = T[keyof T];
 
 /**
- * Versioned storage object.
- *
- * All storages must extend this class to standardize behavior
- */
-export interface VersionedStorage {
-	/**
-	 * Public name for write in versions DB
-	 */
-	publicName: string;
-
-	/**
-	 * Storage version must be increase with every change of storage structure.
-	 *
-	 * While run application this version will compared to version from storages database and if it different, will call method `updateStorageVersion`
-	 */
-	storageVersion: number;
-
-	/**
-	 * Method which call for each update storage version
-	 *
-	 * This method may be used to implement migration data between application versions,
-	 * such as migrations must run only by condition
-	 */
-	updateStorageVersion: (prevVersion: number | null) => Promise<void>;
-}
-
-/**
- * Abstract class for `VersionedStorage` interface
- */
-export abstract class AbstractVersionedStorage {
-	public static storageVersion = 0;
-
-	// public static async updateStorageVersion(_prevVersion: number | null) { }
-}
-
-/**
  * Make all nested properties of object are optional
  */
 export type DeepPartial<T> = T extends object
